Skip test when needed applications are missing
On some test hosts a lot of applications are skipped from the erlang
installation in order to make test go faster.

systools_SUITE:app_start_type_relup uses a few applications which
therefore might not exist. This commit make sure the test is skipped
(instead of fail) if some of the needed applications are missing.