Store expanded config in application_controller:change_application_data
If the Config given to application_controller:change_application_data
included other config files, it was only expanded for already existing
(loaded) applications. If an upgrade added a new application which had
config data in an included config file, the new application did not
get correct config data.

This is now changed so config data will be expanded for all
applications.