BER: Suppress dialyzer warnings for encode_bit_string/4
We don't want to touch the code used for encoding BIT STRINGs when
'legacy_erl_types' is active, since it will be removed within two
or three major releases. But we do want to suppress the dialyzer
warnings in the meantime. The easiest way is to call
encode_bit_string/4 with unknown types from an exported function
that is never actually called like this:

-export(['dialyzer-suppressions'/0]).

'dialyzer-suppressions'(Arg) ->
    {A,B,C,D} = Arg,
    encode_bit_string(A, B, C, D),
    ok.