Test suites: Add support for running Dialyzer on generated code
Running dialyzer on test suites is good for finding both bugs and
useless dialyzer warnings. Add a mechanism in asn1_test_lib for
turning on running of dialyzer on all generated code.

Turned off by default, because it is slow and because there are
still a few dialyzer warnings left.