ssl: implement ssl:getstat/{1,2}
ssl:getstat is similar to inet:getstat on raw socket.
This implementation keeps track of sent and received packet count and
sizes.
Options '*_oct', '*_cnt', '*_max' are updated when packet is sent or
received, '*_avg' ones are calculated when handling a call to getstat,
others silently fall back to zero instead of raising {error, einval}.
This means that 'recv_dvi' and 'send_pend' options are always zero.