Make time axis of Observer graphs adjustable
This adds a dialog box in which the user may set the time axis of the
performance graphs in Observer (in minutes and seconds). To make the
graphs more readable for different time axis lengths, the time markers
for the graphs have been rewritten so that they adjust their interval
depending on time axis length. The labels of the time markers will show
only seconds, only minutes, or both, depending on the interval.