add_abstract_code: Remove 'from_asm' option
The purpose of add_abstract_code is to give Dialyzer some
abstract code so that it will not fail fatally when analysing
prim_eval which was compiled from BEAM assembly.

But if Dialyzer were to pass along the compiler options
that the module was compiled with when translating the
abstract code to Core Erlang, the 'from_asm' option would
crash the compilation. Thus, since we are already cheating,
we should cheat a little bit more and also remove the
'from_asm' option.