kernel,ssh: Add synchronous user_drv protocol
Added a put_chars_sync to the protocol that can be used to
talk to user_drv and made group use it. This is needed in order
to guarantee that bytes has been pushed to the tty port when
doing something like this:

io:format("halting\n"),erlang:halt(0).

Before this change the halting message could be lost in the message
queue of the user_drv process, this is no longer possible.

This commit also fixes ssh_cli as that plugs itself in as a user_drv
process.

OTP-12240