Fix counting of outgoing requests
Commit df19c272 broke this in avoiding counting on arbitrary keys.
It didn't break it sufficiently for the only counters usage in the test
suites to fail however: watchdog counters worked as intended, but no
others, not even CER and DPR. More testcases are needed.

This commit does change/fix the previous semantics somewhat:

- Retransmissions are no longer counted. This previously made it
  impossible to distinguish between these and unanswered requests, since
  both counted as an outgoing request. There should probably be a
  retransmission counter but it should be distinct from the sent request
  counter.

- The counting is always on the node from which diameter:call/4 is
  invoked, not the node on which the transport resides, as was previously
  the case. (Although they're typically one and the same.)

Note that none of these semantics are documented as yet, so we're not
changing a documented interface.