correction of huge status bar
following the last evolution - use wxStatusBar:new(Frame) then wxFrame:setStatusBar(Frame, StatusBar), the status bar use the full space available in frame.
move the creation of the status bar after the creation of the notebook and the mainsizer.

it is ok on windows 7 64bits, no other test done so far.