Correct expansion of parameterized types
Attempting to compile:

SomeType{SOME-CLASS-NAME, SOME-CLASS-NAME:SomeSet} ::= ...
        SEQUENCE {
            something   SOME-CLASS-NAME.&id({SomeSet})
        }

would crash the compiler, because the actual parameter for
SOME-CLASS-NAME was not substituted into the governor for
the SomeSet parameter.

While we are at it, combine the functionality of is_class/2
and get_class_def/2 (eliminating is_class/2). Most callers
call both function.