[ct] Add ct_release_tests.erl
This is a library module for testing release related functionality in
one or more applications. The first version of ct_release_tests
include test of upgrade only.

Note that the hipe application does not support any upgrade, and typer
and dialyzer requires hipe, so these three applications are always
exluded from the test.