[observer] Remove note about R15B nodes in user guide
A note in the user guide says "Only R15B nodes can be observed". This
note was written because the observer_backend module did not exist in
earlier releases. It should have said "R15B or later". However, now
that we are on OTP-17, we don't think the note is necessary anymore so
it is now removed.