Simplify handling of an empty object definition
When the parser sees:

  something SOMETHING ::= {}

it has no way of knowing whether 'something' is an value or an
object. It depends on how SOMETHING is defined.  For example:

  SOMETHING ::= SEQUENCE {}

or

  SOMETHING ::= CLASS { &id OPTIONAL }

Because of that ambiguity, there is no way to avoid a special case
when we check an object definition. However, there is no need to
invent an entire new checking function for this special case.  It is
much easier to just pretend that the parser gave us
{object,defaultsyntax,[]} and let check_objectdefn/3 check it in the
usual way.