Simplify and correct tag handling
There is no reason to handle tags differently depending on the
back-end. The PER back-end will simply ignore tags.

There is also a bug in tags the ABSTRACT-SYNTAX and TYPE-IDENTIFIER
pre-defined classes. So far it has not caused problems, but it could
do in a future commit, such as the next commit...